+++ src/main/java/racingcar/domain/rules/CarMoveRule.java
@@ -1,4 +1,4 @@
-package racingcar.domain.model;
+package racingcar.domain.rules;
 
 public interface CarMoveRule {
     boolean isGo();
